docs: reinstate the ASAP7 view; both physical views now do full place-and-route
Section 3 of the plan revision had replaced the ASAP7 predictive view on the
finding that ASAP7 existed nowhere on this filesystem, openroad was not
installed, and the container the archived results came from was absent. That
reasoning was right about the machine and wrong about the conclusion: the
machine has network access and the missing pieces were fetchable, in about
twenty minutes.

The decisive check is that the archived ASAP7 reduction case was re-run and
reproduced bit-for-bit - every metric identical and a matching 6_final.v
SHA-256 - so the archived campaign is reproducible in fact, not merely in
principle, and is reinstated as evidence.

Both views now do full place-and-route rather than synthesis and static timing
only, and all four block runs closed with zero detailed-route DRC and zero
antenna violations. SKY130 HD is the 130 nm manufacturable foundry view and
ASAP7 the 7 nm predictive one. IHP SG13G2 was rejected for the digital views
only because it is also 130 nm, which would have given two views at one node
and no scaling comparison; its ROM-bitcell SPICE evidence and its
place-and-route result remain valid for the blocks they cover.

The section records how it reached the wrong conclusion, because the failure
mode matters more than the fact: it inferred a permanent constraint from an
observation about the current filesystem, and proposed changing a plan target
to fit a tooling gap, without first testing whether the gap was removable.

The physical worker also fixed the status field properly rather than patching
the two artifacts I flagged - status is now the engineering verdict and
flow_completed separately reports whether the script ran, with a test that
walks every artifact and fails if any claims pass with unmet timing. It caught
all ten stale files. Two further bugs surfaced in the process, including ORFS
slack being recorded in nanoseconds when it was picoseconds, a 1000x error.

+++ tests/compiler/test_rom_backend.py
@@ -224,7 +224,7 @@ def qwen_shaped_graph(
     span = Symbolic("span_tokens", 1, span_max)
     context = Symbolic("context_tokens", 1, span_max)
 
-    token_ids = builder.value("input.token_ids", "i64", (span,), "input")
+    token_ids = builder.value("input.token_ids", "u32", (span,), "input")
     positions = builder.value("input.positions", "i32", (span,), "input")
     embed = builder.weight("model.embed_tokens.weight", "bf16", (vocabulary, hidden))
     residual = builder.value("sequence.embedding", "bf16", (span, hidden), "activation")
@@ -234,7 +234,7 @@ def qwen_shaped_graph(
         (token_ids, embed),
         (residual,),
         contract="bf16_payload_lookup_v1",
-        attributes={"input_dtype": "i64", "output_dtype": "bf16"},
+        attributes={"input_dtype": "u32", "output_dtype": "bf16"},
     )
 
     states = tuple(
@@ -489,7 +489,7 @@ def qwen_shaped_graph(
         (logits,),
         contract="bf16_bf16_fp32_sequential_rne_v1",
     )
-    selected = builder.value("output.selected", "i32", (1,), "activation")
+    selected = builder.value("output.selected", "u32", (1,), "activation")
     builder.kernel(
         "token_selection",
         "ARGMAX",
@@ -498,7 +498,7 @@ def qwen_shaped_graph(
         contract="greedy_lowest_token_id_argmax_v1",
         attributes={"input_dtype": "bf16", "output_dtype": "u32"},
     )
-    next_token = builder.value("output.next_token", "i64", (1,), "output")
+    next_token = builder.value("output.next_token", "u32", (1,), "output")
     builder.kernel(
         "token_append",
         "TOKEN_APPEND",
@@ -549,7 +549,7 @@ def deepseek_shaped_graph(
     context = Symbolic("context_tokens", 1, span_max)
     layers = len(sequence)
 
-    token_ids = builder.value("input.token_ids", "i64", (span,), "input")
+    token_ids = builder.value("input.token_ids", "u32", (span,), "input")
     positions = builder.value("input.positions", "i32", (span,), "input")
     embed = builder.weight("model.embed_tokens.weight", "bf16", (vocabulary, hidden))
     residual = builder.value("sequence.embedding", "bf16", (span, hidden), "activation")
@@ -559,7 +559,7 @@ def deepseek_shaped_graph(
         (token_ids, embed),
         (residual,),
         contract="bf16_payload_lookup_v1",
-        attributes={"input_dtype": "i64", "output_dtype": "bf16"},
+        attributes={"input_dtype": "u32", "output_dtype": "bf16"},
     )
     states = tuple(
         StateResource(
@@ -874,7 +874,7 @@ def deepseek_shaped_graph(
         (logits,),
         contract="bf16_bf16_fp32_sequential_rne_v1",
     )
-    selected = builder.value("output.selected", "i32", (1,), "activation")
+    selected = builder.value("output.selected", "u32", (1,), "activation")
     builder.kernel(
         "token_selection",
         "ARGMAX",
@@ -883,7 +883,7 @@ def deepseek_shaped_graph(
         contract="greedy_lowest_token_id_argmax_v1",
         attributes={"input_dtype": "bf16", "output_dtype": "u32"},
     )
-    next_token = builder.value("output.next_token", "i64", (1,), "output")
+    next_token = builder.value("output.next_token", "u32", (1,), "output")
     builder.kernel(
         "token_append",
         "TOKEN_APPEND",
